working on customer class. customer class holds an arraylist which is their item 'shopping cart'. completed getTotalCost and getOrderNumber methods. also got started on the Order class. order class will have to take in customer for its' methods rather than items.

+++ src/Memory.java
@@ -1,5 +1,5 @@
 public class Memory extends Item{
-    public Memory (String name, Float price, int wattage){ //constructor
+    public Memory (String name, double price, int wattage){ //constructor
         this.setItemName(name);
         this.setItemCost(price);
         this.setWattage(wattage);
